requirements.txt
nicegui>=1.4.0
passlib==1.7.4
bcrypt==3.2.2
pillow>=10.0
matplotlib>=3.7
pandas>=2.0

sqlalchemy>=2.0
psycopg2-binary>=2.9
python-dotenv>=1.0

supabase